// .angular/src/app/models/productoPorSKU.interface.ts
export interface ProductosPorSKU {
    Codigo_cliente:       string;
    Nombre:               string;
    Calle:                string;
    Ciudad:               string;
    Provincia:            string;
    Latitud:              string;
    Longitud:             string;
    Telefono:             string;
    Email:                string;
    Codigo_pedido:        string;
    Fecha_pedido:         Date;
    Operacion:            string;
    Codigo_producto:      string;
    Descripcion_producto: string;
    Cantidad_producto:    string;
    Peso:                 string;
    Volumen:              string;
    Dinero:               string;
    Duracion_min:         string;
    Ventana_horaria_1:    string;
    Ventana_horaria_2:    string;
    Notas:                string;
    Agrupador:            string;
    Email_remitentes:     string;
    Eliminar_pedido:      string;
    Vehiculo:             string;
    Habilidades:          string;
    SKU:                  string;
    Pistoleado:           string;
    Tamaño:               string;
    Estado:               string;
    En_ruta:              string;
    Diferencias_dias:    number;
}